use postgre instead of sqlite to have transactions

+++ api/migrations/1_create-metadata-table.sql
@@ -2,6 +2,6 @@ create table if not exists audio_metadata (
     identifier varchar(512) primary key,
     name varchar(255),
     author varchar(255),
-    duration integer,
+    duration bigint,
     cover_art_url varchar(512)
 );
